+++ CHANGELOG.md
@@ -143,7 +143,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - Initial version.
 
-[unreleased]: https://github.com/extendssoftware/exa-php/compare/0.2.0...HEAD
+[unreleased]: https://github.com/extendssoftware/exa-php/compare/0.3.0...HEAD
 
 [0.3.0]: https://github.com/extendssoftware/exa-php/commits/0.3.0